Azure edge module development/Niyantran/modules/pyworker/requirements.txt
#azure-iot-device~=2.7.0
pytz==2019.3
paho-mqtt
termcolor